+++ lib/authentication/presentation/page/signup_page_state.dart
@@ -11,6 +11,11 @@ import '../widget/password_field_notifier.dart';
 
 part 'signup_page_state.freezed.dart';
 
+// --------------------------------------------------
+//
+//   Freezed
+//
+// --------------------------------------------------
 @freezed
 class SignUpPageState with _$SignUpPageState {
   factory SignUpPageState({
@@ -24,4 +29,33 @@ class SignUpPageState with _$SignUpPageState {
     required Function useAuth,
     @Default("新規登録") String title,
   }) = _SignUpPageState;
+}
+
+// --------------------------------------------------
+//
+//  StateNotifier
+//
+// --------------------------------------------------
+class SignUpPageStateNotifier extends StateNotifier<SignUpPageState> {
+  SignUpPageStateNotifier({
+    required LoginFormState loginFormState,
+    required LoginIdFieldState loginIdFieldState,
+    required LoginIdFieldStateNotifier loginIdFieldStateNotifier,
+    required PasswordFieldState passwordFieldState,
+    required PasswordFieldStateNotifier passwordFieldStateNotifier,
+    required LoginSubmitState loginSubmitState,
+    required Function onSubmit,
+    required Function useAuth,
+  }) : super(
+          SignUpPageState(
+            loginFormState: loginFormState,
+            loginIdFieldState: loginIdFieldState,
+            loginIdFieldStateNotifier: loginIdFieldStateNotifier,
+            passwordFieldState: passwordFieldState,
+            passwordFieldStateNotifier: passwordFieldStateNotifier,
+            loginSubmitState: loginSubmitState,
+            onSubmit: onSubmit,
+            useAuth: useAuth,
+          ),
+        );
 }
